Introduce new MemoryRecording for use with Jupyter notebooks
